Add Cancancan Authorizations

+++ app/controllers/api/v1/cars_controller.rb
@@ -1,56 +1,82 @@
 class Api::V1::CarsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
   before_action :find_car, only: %i[show update destroy]
+
   # GET /cars
   def index
-    @cars = Car.all
-    if @cars.empty?
-      render_response(:not_found)
+    if can? :read, Car
+      @cars = Car.all
+      if @cars.empty?
+        render_response(:not_found)
+      else
+        render_response(:found, payload: @cars)
+      end
     else
-      render_response(:found, payload: @cars)
+      render_response(:unauthorized)
     end
   end
 
   # GET /cars/:id
   def show
-    if @car.nil?
-      render_response(:not_found)
+    if can? :read, Car
+      if @car.nil?
+        render_response(:not_found)
+      else
+        render_response(:found, payload: @car)
+      end
     else
-      render_response(:found, payload: @car)
+      render_response(:unauthorized)
     end
   end
 
   # POST /cars
   def create
-    @car = Car.new(car_params)
-    if @car.save
-      render_response(:created)
+    if can? :create, Car
+      @car = Car.new(car_params)
+      if @car.save
+        render_response(:created)
+      else
+        render_response(:unable_to_create)
+      end
     else
-      render_response(:unable_to_create)
+      render_response(:unauthorized)
     end
   end
 
   # PUT /cars/:id
   def update
-    if %i[model image description price brand year color is_electric accidents].any? { |param| params[param].present? }
-      @car.update(car_params) ? render_response(:updated) : render_response(:unable_to_update)
+    if can? :update, Car
+      if car_params.blank? || car_params.empty? || car_params.nil?
+        render_response(:none_attribute)
+      else
+        @car.update(car_params) ? render_response(:updated) : render_response(:unable_to_update)
+      end
     else
-      render_response(:none_attribute)
+      render_response(:unauthorized)
     end
   end
 
   # DELETE /cars/:id
   def destroy
-    if @car.destroy
-      render_response(:deleted)
+    if can? :destroy, Car
+      if @car.destroy
+        render_response(:deleted)
+      else
+        render_response(:unable_to_delete)
+      end
     else
-      render_response(:unable_to_delete)
+      render_response(:unauthorized)
     end
   end
 
   private
 
   def car_params
+    # Accessing a parameter that doesn't exist will raise a KeyError exception
     params.require(:car).permit(:model, :image, :description, :price, :brand, :year, :color, :is_electric, :accidents)
+  rescue KeyError
+    # Handle the exception by setting a default value for the missing parameter
+    {}
   end
 
   def find_car
